modified methods of user added views for listing categories andproducts, changed returning data from the methods 'add to cart', 'del from cart', 'cart to order', refurbished method cart to order

+++ shop/models.py
@@ -5,18 +5,20 @@ from django.core.exceptions import ObjectDoesNotExist
 from datetime import datetime, timedelta
 from django.db.models import Sum, F, ExpressionWrapper
 from django.contrib.auth.decorators import login_required
-from shop.exceptions import QuantityKeyError
+from shop.exceptions import (QuantityKeyError, CartItemIdKeyError)
 
 
 class Category(models.Model):
     name = models.CharField(max_length=64, unique=True)
     parent = models.ForeignKey('self', null=True, max_length=64, verbose_name='Parent', blank=True,
                                db_index=True, on_delete=models.CASCADE)
-    icon = models.ImageField(upload_to='', verbose_name=u'Photo', help_text='jpg/png - file')
+    photo = models.ImageField(upload_to='', verbose_name=u'Photo', help_text='jpg/png - file')
     description = models.TextField(max_length=512, blank=True, null=True)
+
 
     def __str__(self):
         return self.name
+
 
 
 class Product(models.Model):
@@ -24,8 +26,8 @@ class Product(models.Model):
     name = models.CharField(max_length=64, db_index=True)
     category = models.ForeignKey(Category, max_length=64, verbose_name='Category', db_index=True,
                                  on_delete=models.CASCADE)
-    price = models.DecimalField(null=False, decimal_places=2, max_digits=7)
-    discount = models.DecimalField(default=0, decimal_places=2, max_digits=7)
+    price = models.DecimalField(null=False, decimal_places=2, max_digits=9)
+    discount = models.DecimalField(default=0, decimal_places=2, max_digits=9)
     qu_in_stock = models.PositiveIntegerField(default=0, db_index=True)
     reserved = models.PositiveIntegerField(default=0, blank=True, db_index=True)
     allow_res_per_user = models.SmallIntegerField(default=1)
@@ -39,12 +41,11 @@ class Product(models.Model):
 
 
 class User(AbstractUser):
-    # full_name = models.CharField(max_length=128, db_index=True, null=False)
+    
     phone = models.CharField(max_length=13, unique=True, db_index=True)
     address = models.CharField(max_length=256, blank=True, null=True)
     zip_code = models.IntegerField(null=True, blank=True)
     birth_day = models.DateField(default=None, null=True)
-    # purse = models.DecimalField(default=0, decimal_places=2, max_digits=10)
 
     # EMAIL_FIELD = 'email'
     # USERNAME_FIELD = 'username'
@@ -84,7 +85,7 @@ class User(AbstractUser):
             product = Product.objects.get(code=product_code)
             reserved = Reserve.objects.get(product=product, user=self)
         except ObjectDoesNotExist:
-            return 0
+            return 0, 0
         return reserved.quantity, reserved
 
     @transaction.atomic
@@ -107,13 +108,16 @@ class User(AbstractUser):
         cart, jc = self.get_cart()
         return jc
 
-    def del_from_cart(self, *ids):
-            cart = self.get_cart()
-            deleted = cart.filter(id__in=ids).delete()
-            return deleted
+    def del_from_cart(self, ids):
+            cart, jc = self.get_cart()
+            deleted, d = cart.filter(id__in=ids).delete()
+            cart, jc = self.get_cart()
+            if not deleted:
+                raise CartItemIdKeyError
+            return jc
 
     def cart_to_order(self):  # save products from cart to order
-        cart = self.get_cart()
+        cart, jc = self.get_cart()
         order_items = []
         self.del_old_reserves()
         if cart:
@@ -122,27 +126,33 @@ class User(AbstractUser):
 
                 for obj in cart:
                     res_quantity, res_item = self.check_reserved(obj.product.code)
-                    if res_quantity >= obj.quantity:
-                        res_item.quantity -= obj.quantity
-                        res_item.save()
-                        obj.product.reserved -= obj.quantity
-                        obj.product.save()
-                    else:
-                        obj.product.reserved -= res_quantity
-                        res_item.delete()
+                    if res_quantity:
+                        if res_quantity >= obj.quantity:
+                            res_item.quantity -= obj.quantity
+                            res_item.save()
+                            obj.product.reserved -= obj.quantity
+                            obj.product.save()
+                        else:
+                            obj.product.reserved -= res_quantity
+                            res_item.delete()
                     if (obj.product.qu_in_stock - obj.product.reserved - obj.quantity) >= 0:
                         order_items.append(Order_items(order=order, code=obj.product.code, name=obj.product.name,
-                                                       price=obj.product.price, discount=obj.product.discount))
+                                                           quantity=obj.quantity,
+                                                           price=obj.product.price, discount=obj.product.discount))
+                        obj.product.qu_in_stock -= obj.quantity
+                        obj.product.save()
                     else:
                         raise ValueError(f"Not enough quantity of product {obj.product} "
-                                        f"in stock. You should to choose a smaller quantity!")
+                                            f"in stock. You should to choose a smaller quantity!")
                 Order_items.objects.bulk_create(order_items)
                 cart_wt = Order_items.objects.filter(order=order).aggregate(cart_total=ExpressionWrapper(
                     Sum(F('price') * F('quantity')), output_field=models.DecimalField()))
                 order.total_cost = cart_wt['cart_total']
                 order.save()
-                deleted = cart.delete()
-                result = {'result': 'success', 'reason': f'replaced {deleted[0]}'}
+                cart.delete()
+                order_val = Order.objects.values().get(id=order.id)
+                items = list(order.order_items_set.values())
+                result = {'order': order_val, 'order_items': items}
         return result
 
     @login_required
@@ -184,7 +194,7 @@ class Cart_items(models.Model):
 
 class Order(models.Model):
     user = models.ForeignKey(User, related_name='orders', on_delete=models.CASCADE)
-    total_cost = models.DecimalField(default=0, decimal_places=2, max_digits=8)
+    total_cost = models.DecimalField(default=0, decimal_places=2, max_digits=8, null=True)
     date = models.DateTimeField(auto_now_add=True)
     paid = models.BooleanField(default=False, choices=(('True', True), ('False', False)))
     status = models.CharField(max_length=16, default='processing')
@@ -194,13 +204,14 @@ class Order_items(models.Model):
     order = models.ForeignKey(Order, on_delete=models.CASCADE)
     code = models.CharField(max_length=32)
     name = models.CharField(max_length=64)
-    price = models.DecimalField(null=False, decimal_places=2, max_digits=7)
-    discount = models.FloatField()
+    quantity = models.PositiveIntegerField(default=1)
+    price = models.DecimalField(null=False, decimal_places=2, max_digits=9)
+    discount = models.DecimalField(null=False, decimal_places=2, max_digits=9)
 
 
 class Reserve(models.Model):
-    product = models.ForeignKey(Product, on_delete=models.deletion.CASCADE)
-    user = models.ForeignKey(User, on_delete=models.deletion.CASCADE)
+    product = models.ForeignKey(Product, on_delete=models.CASCADE)
+    user = models.ForeignKey(User, on_delete=models.CASCADE)
     quantity = models.PositiveIntegerField(default=1)
     datetime_of_reserve = models.DateTimeField(auto_now_add=True)
